[WIP] Start moving stuff to llvm-args
