+++ kani-compiler/kani_queries/src/lib.rs
@@ -1,9 +1,8 @@
 // Copyright Kani Contributors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use std::sync::atomic::{AtomicBool, Ordering};
 #[cfg(not(feature = "unsound_experiments"))]
-use std::sync::Mutex;
+use std::sync::{Arc, Mutex};
 use strum_macros::{AsRefStr, EnumString, EnumVariantNames};
 
 #[cfg(feature = "unsound_experiments")]
@@ -56,60 +55,76 @@ pub trait UserInput {
     fn get_stubbing_enabled(&self) -> bool;
 
     #[cfg(feature = "unsound_experiments")]
-    fn get_unsound_experiments(&self) -> Arc<Mutex<UnsoundExperiments>>;
+    fn get_unsound_experiments(&self) -> UnsoundExperiments;
 }
 
-#[derive(Debug, Default)]
+/// This structure should only be used behind a synchronized reference or a snapshot.
+#[derive(Debug, Clone)]
 pub struct QueryDb {
-    check_assertion_reachability: AtomicBool,
-    emit_vtable_restrictions: AtomicBool,
-    json_pretty_print: AtomicBool,
-    ignore_global_asm: AtomicBool,
-    reachability_analysis: Mutex<ReachabilityType>,
+    check_assertion_reachability: bool,
+    emit_vtable_restrictions: bool,
+    json_pretty_print: bool,
+    ignore_global_asm: bool,
+    reachability_analysis: ReachabilityType,
     stubbing_enabled: bool,
     #[cfg(feature = "unsound_experiments")]
-    unsound_experiments: Arc<Mutex<UnsoundExperiments>>,
+    unsound_experiments: UnsoundExperiments,
+}
+
+impl QueryDb {
+    pub fn new() -> Arc<Mutex<QueryDb>> {
+        Arc::new(Mutex::new(QueryDb {
+            check_assertion_reachability: false,
+            emit_vtable_restrictions: false,
+            json_pretty_print: false,
+            ignore_global_asm: false,
+            reachability_analysis: ReachabilityType::None,
+            stubbing_enabled: false,
+            #[cfg(feature = "unsound_experiments")]
+            unsound_experiments: false,
+        }))
+    }
 }
 
 impl UserInput for QueryDb {
     fn set_emit_vtable_restrictions(&mut self, restrictions: bool) {
-        self.emit_vtable_restrictions.store(restrictions, Ordering::Relaxed);
+        self.emit_vtable_restrictions = restrictions;
     }
 
     fn get_emit_vtable_restrictions(&self) -> bool {
-        self.emit_vtable_restrictions.load(Ordering::Relaxed)
+        self.emit_vtable_restrictions
     }
 
     fn set_check_assertion_reachability(&mut self, reachability: bool) {
-        self.check_assertion_reachability.store(reachability, Ordering::Relaxed);
+        self.check_assertion_reachability = reachability;
     }
 
     fn get_check_assertion_reachability(&self) -> bool {
-        self.check_assertion_reachability.load(Ordering::Relaxed)
+        self.check_assertion_reachability
     }
 
     fn set_output_pretty_json(&mut self, pretty_json: bool) {
-        self.json_pretty_print.store(pretty_json, Ordering::Relaxed);
+        self.json_pretty_print = pretty_json;
     }
 
     fn get_output_pretty_json(&self) -> bool {
-        self.json_pretty_print.load(Ordering::Relaxed)
+        self.json_pretty_print
     }
 
     fn set_ignore_global_asm(&mut self, global_asm: bool) {
-        self.ignore_global_asm.store(global_asm, Ordering::Relaxed);
+        self.ignore_global_asm = global_asm;
     }
 
     fn get_ignore_global_asm(&self) -> bool {
-        self.ignore_global_asm.load(Ordering::Relaxed)
+        self.ignore_global_asm
     }
 
     fn set_reachability_analysis(&mut self, reachability: ReachabilityType) {
-        *self.reachability_analysis.get_mut().unwrap() = reachability;
+        self.reachability_analysis = reachability;
     }
 
     fn get_reachability_analysis(&self) -> ReachabilityType {
-        *self.reachability_analysis.lock().unwrap()
+        self.reachability_analysis
     }
 
     fn set_stubbing_enabled(&mut self, stubbing_enabled: bool) {
